refactor stores with modal helper
