+++ src/stores/captureLimitModal.ts
@@ -1,20 +1,14 @@
 import { defineStore } from 'pinia'
 import { ref } from 'vue'
-import { useMobileTabStore } from './mobileTab'
+import { createModalStore } from './helpers'
 
 export const useCaptureLimitModalStore = defineStore('captureLimitModal', () => {
-  const isVisible = ref(false)
+  const { isVisible, open: openModal, close } = createModalStore('game')
   const requiredLevel = ref(0)
-  const mobile = useMobileTabStore()
 
   function open(level: number) {
     requiredLevel.value = level
-    mobile.set('game')
-    isVisible.value = true
-  }
-
-  function close() {
-    isVisible.value = false
+    openModal()
   }
 
   return { isVisible, requiredLevel, open, close }
